# specs/001-intellistack-platform/contracts/learning.openapi.yaml
openapi: 3.1.0
info:
  title: IntelliStack Learning API
  version: 1.0.0
  description: Learning progress, prerequisites, badges, certificates (FR-001 to FR-015)
servers:
  - url: /api/v1/learning
    description: Learning API base path
...
paths:
...
  /challenge/assessment/{stageId}:
    get:
      tags: [Challenge]
      summary: Get challenge assessment for stage
      operationId: getChallengeAssessment
      security:
        - sessionAuth: []
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Challenge assessment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeAssessmentResponse'
        '403':
          description: Not eligible for challenge pathway
    post:
      tags: [Challenge]
      summary: Submit challenge assessment
      operationId: submitChallengeAssessment
      security:
        - sessionAuth: []
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeSubmission'
      responses:
        '200':
          description: Challenge result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResultResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    ChallengeAssessmentResponse:
      type: object
      properties:
        stageId:
          type: string
          format: uuid
        assessmentType:
          type: string
          enum: [comprehensive_exam, portfolio_review, live_demonstration]
        instructions:
          type: string
        competencies:
          type: array
          items:
            type: string
        passingThreshold:
          type: number
          format: float
          description: 0.85 (85%)
    ChallengeSubmission:
      type: object
      properties:
        assessmentType:
          type: string
          enum: [comprehensive_exam, portfolio_review]
        responses:
          type: object
        portfolioUrls:
          type: array
          items:
            type: string
            format: uri
    ChallengeResultResponse:
      type: object
      properties:
        passed:
          type: boolean
        score:
          type: number
          format: float
        competenciesMet:
          type: array
          items:
            type: string
        competenciesGaps:
          type: array
          items:
            type: string
        stageSkipped:
          type: boolean
        safetyRefresherRequired:
          type: boolean
...